internal manipulation interface test
