+++ kubernete_index.go
@@ -3,7 +3,7 @@ package lambda
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/discovery"
 )
 
 var indexerInitialized bool
@@ -22,24 +22,24 @@ type resourceIndexerImpl struct {
 	store map[Resource]metav1.APIResource
 }
 
-func initIndexer(clientset kubernetes.Interface) {
+func initIndexer(i discovery.DiscoveryInterface) {
 	indexer := &resourceIndexerImpl{
 		store: make(map[Resource]metav1.APIResource),
 	}
-	grouplist, err := clientset.Discovery().ServerGroups()
+	grouplist, err := i.ServerGroups()
 	if err != nil {
 		panic(err)
 	}
 	for _, g := range grouplist.Groups {
 		for _, v := range g.Versions {
-			rs, err := clientset.Discovery().ServerResourcesForGroupVersion(v.GroupVersion)
+			rs, err := i.ServerResourcesForGroupVersion(v.GroupVersion)
 			if err != nil {
 				panic(err)
 			}
 			for _, r := range rs.APIResources {
 				for _, supportedResource := range GetResources() {
 					supportedResource := supportedResource
-					if supportedResource.GetCanonicalName() == r.Name {
+					if string(supportedResource) == r.Kind {
 						r.Group = g.Name
 						r.Version = v.Version
 						indexer.store[supportedResource] = r
