+++ kubernetes_index_test.go
@@ -6,8 +6,11 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestSimpleClientSetDiscovery(t *testing.T) {
+func init() {
 	OutOfClusterDefault()
+}
+
+func TestSimpleClientSetDiscovery(t *testing.T) {
 	for _, resource := range GetResources() {
 		_, exist := indexerInstance.(*resourceIndexerImpl).store[resource]
 		assert.NotEmpty(t, exist, "resource %s index missing", resource)
